[Genlist] Fix performance issue of compress mode
Before, genlist items are always unrealized/realized when genlist is
resized if compress mode is on.
But those needs only when item's height is changed.
Above case is that texblock is used and ellipsis is not applied for the
textblock. In that case, if genlist is resized, lines of each textblock
part can be changed and also item's height can be changed.
So after _elm_genlist_pan_smart_resize() is called, size_changed is
checked. And, when the calc_job()is called, _changed_job is called for
checking if height is changed.
Also if genlist viewport is not initialized, it is not needed to run
_queue_process(), because edje min calc of textblock can not performed
correctly and textblock's height can be larger if viewport's width is 0
and word wrap (not ellpsis) is used.

Change-Id: I3d6bdffaee3a3a7a95bde6bbd1a6118764b1fe0e